fix(cannot buy),update(cs#),fix(player invalid)
fix(cannot buy): cannot buy after infinite grenades when player join
fix(player invalid): player or pawn may be invalid and cause exception.
update(cs#): 239 to 242

+++ FunHealTeammates.cs
@@ -80,7 +80,10 @@ public class FunHealTeammates : FunBaseClass
             
             if (Enabled == false) return HookResult.Stop;
             Timer ?playerTimer;
-            CCSPlayerPawn ?pawn = @event.Userid!.OriginalControllerOfCurrentPawn.Get()!.PlayerPawn.Get();
+            if (!@event.Userid!.IsValid) return HookResult.Continue;
+            var oringin = @event.Userid!.OriginalControllerOfCurrentPawn.Get()!;
+            if (oringin is null) return HookResult.Continue;
+            CCSPlayerPawn ?pawn = oringin.PlayerPawn.Get();
             if (pawn is null) return HookResult.Continue;
             playerTimer = plugin.AddTimer(BurnAfterSecond,() => BurnPlayer(pawn!),TimerFlags.REPEAT);
             playerTimersDict.TryAdd((int)@event.Userid.UserId!,playerTimer);
@@ -92,6 +95,7 @@ public class FunHealTeammates : FunBaseClass
             
             if (Enabled == false) return HookResult.Stop;
             Timer ?playerTimer;
+            if (!@event.Userid!.IsValid) return HookResult.Continue;
             playerTimersDict.TryGetValue((int)@event.Userid!.UserId!,out playerTimer);
             if (playerTimer is null) return HookResult.Continue;
             playerTimer.Kill();
